requirements.txt
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.2
cryptography==37.0.4
Flask==1.1.2
Flask-API==2.0
fuzzywuzzy==0.17.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
nameparser==1.0.6
nltk==3.4.5
numpy==1.18.2
pdfminer.six==20220524
pdfplumber==0.7.4
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.9.4
python-Levenshtein==0.12.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.16.0
sklearn==0.0
sortedcontainers==2.1.0
tabulate==0.8.6
tinydb==3.15.2
typing-extensions==4.3.0
unicodecsv==0.14.1
Wand==0.6.8
Werkzeug==1.0.1
